rename "metafeed" to "api"
